Handle negative amperage case corretly
Amperage in MSP_ANALOG is signed 16-bit integer (unless MultiWii mode
is selected). Handle this correctly.
Proper rounding instead of simple truncation by integer division is
also applied.

+++ MW_OSD/Serial.ino
@@ -298,7 +298,7 @@ void serialMSPCheck()
     MwVBat=read8();
     pMeterSum=read16();
     MwRssi = read16();
-    MWAmperage = read16();
+    MWAmperage = (int16_t)read16();
  }
 
 #ifdef MENU_SERVO
